+++ lib/app/server/rfi-pub-sub-engine.d.ts
@@ -6,6 +6,9 @@ import { ModelDelta, NotificationOf, RfiSubscriptionOptions } from '../../gql';
 import { RfiPubSubConfig } from '../config';
 import { DbModelAndOidScope } from './init-sequelize';
 import { RfiPubSubEngine } from './rfi-pub-sub-engine.interface';
+/**
+ * @NOTE THIS IS IS ONLY FOR CLIENT SUBSCRIPTIONS
+ */
 export declare class RfiPubSub extends GooglePubSub implements RfiPubSubEngine {
     protected topicPrefix: string;
     publisher_version: string;